refactor constraints and test constants
