+++ domain/src/main/java/com/example/domain/museum/Article.java
@@ -16,8 +16,6 @@ import java.util.LinkedHashSet;
 import java.util.Objects;
 import java.util.Set;
 
-import static com.example.constraints.domain.museum.ArticleConstraints.*;
-
 /**
  * Article is domain model of the web-page article business object.
  *
