+++ domain/src/main/java/com/example/domain/museum/Event.java
@@ -15,8 +15,6 @@ import org.springframework.validation.annotation.Validated;
 import java.time.LocalDateTime;
 import java.util.Objects;
 
-import static com.example.constraints.domain.museum.EventConstraints.*;
-
 /**
  * Event is domain model of the museum event business object.
  *
