+++ domain/src/main/java/com/example/domain/users/User.java
@@ -12,8 +12,6 @@ import java.util.LinkedHashSet;
 import java.util.Objects;
 import java.util.Set;
 
-import static com.example.constraints.domain.users.UserConstraints.*;
-
 /**
  * Entity class representing a user in the application.
  *
