+++ libraries/src/main/java/com/example/constraints/SharedConstraints.java
@@ -1,4 +1,4 @@
-package com.example.constraints.domain;
+package com.example.constraints;
 
 public class SharedConstraints {
     private SharedConstraints() {
